Move client to its own file
